Moved roll class to it's own header.
